refactor: move email-validator-adapter file for email-adapter folder
