+++ globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoUri":"mongodb://127.0.0.1:35919/","mongoDBName":"jest"}
+{"mongoUri":"mongodb://127.0.0.1:44965/","mongoDBName":"jest"}
