+++ src/main/adapters/validation/email-validator/email-validator-adapter.ts
@@ -1,5 +1,5 @@
 import validator from 'validator';
-import { EmailValidator } from '../../../presentation/protocols/email-validator';
+import { EmailValidator } from '../../../../presentation/protocols/email-validator';
 export class EmailValidatorAdapter implements EmailValidator {
   validate(email: string): boolean {
     const isValid = validator.isEmail(email);
