+++ src/main/factories/login/login-validation-controller.ts
@@ -2,7 +2,7 @@ import { Validation } from '../../../presentation/protocols/Validation';
 import { EmailValidation } from '../../../validation/email-validation';
 import { RequiredFieldsValidation } from '../../../validation/required-fields-validation';
 import { ValidationComposite } from '../../../validation/validation-composite';
-import { EmailValidatorAdapter } from '../../adapters/validation/email-validator-adapter';
+import { EmailValidatorAdapter } from '../../adapters/validation/email-validator/email-validator-adapter';
 
 export const makeLoginValidation = (): Validation => {
   const fields = ['email', 'password'];
