+++ src/main/factories/signup/signup-validation.ts
@@ -3,7 +3,7 @@ import { CompareFieldsValidation } from '../../../validation/compare-fields-vali
 import { EmailValidation } from '../../../validation/email-validation';
 import { RequiredFieldsValidation } from '../../../validation/required-fields-validation';
 import { ValidationComposite } from '../../../validation/validation-composite';
-import { EmailValidatorAdapter } from '../../adapters/validation/email-validator-adapter';
+import { EmailValidatorAdapter } from '../../adapters/validation/email-validator/email-validator-adapter';
 
 export const makeSignUpValidation = (): Validation => {
   const fields = ['name', 'email', 'password', 'passwordConfirmation'];
